requirements.txt
Flask==0.12.1
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
Werkzeug==0.12.1
PyJWT==1.4.0
nltk==3.2.2
textblob==0.12.0